+++ src/Text/Gigaparsec/Internal/Errors/ParseError.hs
@@ -38,22 +38,21 @@ data ParseError = VanillaError { presentationOffset :: {-# UNPACK #-} !Word
 {-# INLINABLE fromParseError #-}
 fromParseError :: forall err. ErrorBuilder err => Maybe FilePath -> String -> ParseError -> err
 fromParseError srcFile input err =
-  Builder.format (Builder.pos @err (line err) (col err)) (Builder.source @err srcFile)
-                 (formatErr err)
-  where formatErr :: ParseError -> Builder.ErrorInfoLines err
-        formatErr VanillaError{..} =
+  Builder.build (Builder.pos @err (line err) (col err)) (Builder.source @err srcFile) (buildErr err)
+  where buildErr :: ParseError -> Builder.ErrorInfoLines err
+        buildErr VanillaError{..} =
           Builder.vanillaError @err
             (Builder.unexpected @err (either (const Nothing) (Just . fst) unexpectedTok))
             (Builder.expected @err (Builder.combineExpectedItems @err (Set.map expectItem expecteds)))
             (Builder.combineMessages @err (Set.foldr (\r -> (Builder.reason @err r :)) [] reasons))
-            (Builder.lineInfo @err curLine linesBefore linesAfter caret (trimToLine caretSize))
+            (Builder.lineInfo @err curLine linesBefore linesAfter line caret (trimToLine caretSize))
           where unexpectedTok = unexpectItem lexicalError <$> unexpected
                 caretSize = either id snd unexpectedTok
 
-        formatErr SpecialisedError{..} =
+        buildErr SpecialisedError{..} =
           Builder.specialisedError @err
             (Builder.combineMessages @err (map (Builder.message @err) msgs))
-            (Builder.lineInfo @err curLine linesBefore linesAfter caret (trimToLine caretWidth))
+            (Builder.lineInfo @err curLine linesBefore linesAfter line caret (trimToLine caretWidth))
 
         expectItem :: ExpectItem -> Builder.Item err
         expectItem (ExpectRaw t) = Builder.raw @err t