+++ test/Text/Gigaparsec/Internal/TestError.hs
@@ -17,7 +17,7 @@ data TestErrorLines = VanillaError !(Maybe TestErrorItem) !(Set TestErrorItem) !
 data TestErrorItem = Raw !String | Named !String | EndOfInput deriving stock (Eq, Show, Ord)
 
 instance ErrorBuilder TestError where
-  format p _ = TestError p
+  build p _ = TestError p
 
   type Position TestError = (Word, Word)
   pos = (,)
@@ -45,7 +45,7 @@ instance ErrorBuilder TestError where
   message = id
 
   type LineInfo TestError = Word
-  lineInfo _ _ _ _ width = width
+  lineInfo _ _ _ _ _ width = width
 
   numLinesBefore = 2
   numLinesAfter = 2